clean: AI選択のコンポーネントをリファクタ
